+++ src/Domain/Contract/PluginInterface.php
@@ -4,13 +4,12 @@ declare(strict_types=1);
 
 namespace Iriven\PhpFormGenerator\Domain\Contract;
 
-use Iriven\PhpFormGenerator\Infrastructure\Extension\ExtensionRegistry;
+use Iriven\PhpFormGenerator\Infrastructure\Registry\PluginRegistry;
 
+/**
+ * @api
+ */
 interface PluginInterface
 {
-    public function registerFieldTypes(FieldTypeRegistryInterface $registry): void;
-
-    public function registerFormTypes(FormTypeRegistryInterface $registry): void;
-
-    public function registerExtensions(ExtensionRegistry $registry): void;
+    public function register(PluginRegistry $registry): void;
 }